pyproject.toml
[tool.poetry]
name = "astromap"
version = "0.1.0"
description = "map of astromutual constellations"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.1.3"
pycairo = "^1.27.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
mypy = "^1.13.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
exclude = ["./venv"]
ignore_missing_imports = true

[tool.black]
line-length = 80
target-version = ["py312"]

[tool.isort]
line_length = 80
profile = "black"
